predemo: hub V1 urls + telegram chat-id wiring + demo resolver tests
- ykp-hub: point all 6 modules at live V1 *.oseedigital.tech subdomains (owner/hr/finance/warehouse/investor/ops); narrow ROLE_SSO_APPS to ops (finance-v1 + owner-v1 are POST-only); clear SSO_ROLE
- ykp-finance-v1: telegram_owner_chat_id app_settings row set to presenter chat id 5694784154; TELEGRAM_CHAT_ID env fallback set on VPS
- ykp-warehouse-v1: users U-001 telegram_id set to 5694784154; TELEGRAM_CHAT_ID env set on VPS
- ykp-investor-v1: TELEGRAM_CHAT_ID env set to 5694784154 on VPS
- tests: add demo-chat-id.test.ts per app driving real resolver (getOwnerChatId/resolveRecipients/sendTelegram) against seeded demo row; finance 5/5, warehouse 7/7, investor 3/3 green
- VPS: restart ykp-finance-v1/ykp-warehouse/ykp-investor/ykp-hub PM2; health ok

+++ ykp-hr-v1/src/app/api/hr/adjustments/route.ts
@@ -11,12 +11,12 @@ const insertSchema = z.object({
   employee_id: z.string().min(1),
   adjustment_type: z.enum(['BONUS', 'PENALTY', 'OVERTIME', 'ALLOWANCE', 'CASH_ADVANCE', 'REIMBURSEMENT', 'OTHER']),
   date: z.string().regex(/^\d{4}-\d{2}-\d{2}$/),
-  amount: z.coerce.number().min(0),
+  amount: z.coerce.number().min(0).int(),
   reason: z.string().default(''),
   payroll_period: z.string().regex(/^\d{4}-\d{2}$/),
   // Revisi item 13 — bukti/attachment wajib untuk REIMBURSEMENT & recommended for others
   attachment_url: z.string().default(''),
-  quantity: z.coerce.number().min(0).default(1),
+  quantity: z.coerce.number().min(0).int().default(1),
   unit: z.string().default('')
 });
